add tests for removing a friend | clean urls | remove_friend view

+++ py_backend/friends/views.py
@@ -11,27 +11,44 @@ from django.contrib.auth.decorators import login_required
 
 @login_required
 @require_http_methods(["POST"])
-def send_friend_request(request, user_id):
+def send_request(request, user_id):
     from_user = request.user
     to_user = CustomUser.objects.get(id=user_id)
-    friend_request, created = FriendRequest.objects.get_or_create(
+    interaction_request, created = FriendRequest.objects.get_or_create(
         from_user=from_user,
         to_user=to_user)
     if created:
-        response_data = {'id': friend_request.id, 'status': 'success'}
+        response_data = {'id': interaction_request.id, 'status': 'success'}
         return JsonResponse(response_data, status=200)
     return JsonResponse({'status': 'error'}, status=400)
 
 
 @login_required
 @require_http_methods(["POST"])
-def accept_friend_request(request, request_id):
-	friend_request = FriendRequest.objects.get(id=request_id)
+def accept_friend(request, request_id):
+    friend_request = FriendRequest.objects.get(id=request_id)
+    already_friends = friend_request.to_user.friends.filter(id=friend_request.from_user.id).exists() or friend_request.from_user.friends.filter(id=friend_request.to_user.id).exists()
+    
+    if not already_friends:
+        if friend_request.to_user != request.user:
+            friend_request.to_user.friends.add(friend_request.from_user)
+            friend_request.from_user.friends.add(friend_request.to_user)
+            friend_request.delete()
+            return JsonResponse({'status': 'success'}, status=200)
+        else:
+            return JsonResponse({'status': 'error'}, status=400)
+    else:
+        return JsonResponse({'status': 'Users are already friends'}, status=400)
 
-	if friend_request.to_user != request.user:
-		friend_request.to_user.friends.add(friend_request.from_user)
-		friend_request.from_user.friends.add(friend_request.to_user)
-		friend_request.delete()
-		return JsonResponse({'status': 'success'}, status=200)
-	else:
-		return JsonResponse({'status': 'error'}, status=400)
+@login_required
+@require_http_methods(["POST"])
+def remove_friend(request, request_id):
+    remove_request = FriendRequest.objects.get(id=request_id)
+    are_friends = remove_request.to_user.friends.filter(id=remove_request.from_user.id).exists() or remove_request.from_user.friends.filter(id=remove_request.to_user.id).exists()
+
+    if are_friends:
+        remove_request.from_user.friends.remove(remove_request.to_user)
+        remove_request.to_user.friends.remove(remove_request.from_user)
+        return JsonResponse({'status': 'success'}, status=200)
+    else:
+        return JsonResponse({'status': 'Users are not friends'}, status=400)
